# Conflicts:
#	src/main/java/com/github/meistersky/telegram/command/CallCommand.java

+++ src/main/resources/db/migration/V00001__add_tg_user_table.sql
@@ -4,14 +4,14 @@ DROP TABLE IF EXISTS tg_user;
 
 -- Create tg_user table
 CREATE TABLE tg_user (
-    chat_id INT PRIMARY KEY,
+    chat_id BIGINT PRIMARY KEY,
     active BOOLEAN
 );
 
 -- Create tg_group table
 CREATE TABLE user_group (
     id INT PRIMARY KEY AUTO_INCREMENT,
-    chat_id INT NOT NULL,
+    chat_id BIGINT NOT NULL,
     title VARCHAR(100),
     users VARCHAR(100)
 );
